wait debug edit record

+++ src/main/java/com/example/demo/service/RecordService.java
@@ -1,12 +1,16 @@
 package com.example.demo.service;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +18,7 @@ import com.example.demo.bean.AccountBean;
 import com.example.demo.bean.CategoryBean;
 import com.example.demo.bean.RecordBean;
 import com.example.demo.bean.UserBean;
+import com.example.demo.dto.PageRecordResponse;
 import com.example.demo.dto.RecordRequest;
 import com.example.demo.dto.ValidationResult;
 import com.example.demo.enums.AccountStatus;
@@ -53,14 +58,14 @@ public class RecordService {
 
         ValidationUtils.checkNull(errors, "account_id", accountId, "請選擇帳戶");
 
-        ValidationUtils.checkNull(errors,"category_id",request.getCategory_id(), "請選擇類別");
+        ValidationUtils.checkNull(errors,"category_id",request.getCategoryId(), "請選擇類別");
 
-        ValidationUtils.checkNull(errors, "item_date", request.getItem_date(), "請選擇日期");
+        ValidationUtils.checkNull(errors, "itemDate", request.getItemDate(), "請選擇日期");
 
-        ValidationUtils.checkNull(errors, "item_price", request.getItem_price(), "請輸入金額");
+        ValidationUtils.checkNull(errors, "itemPrice", request.getItemPrice(), "請輸入金額");
 
         
-        ValidationUtils.checkIsNegative(errors, "item_price", request.getItem_price(), "金額不能小於0");
+        ValidationUtils.checkIsNegative(errors, "itemPrice", request.getItemPrice(), "金額不能小於0");
 
         if (errors.isEmpty()) {
             result.setErrors(Optional.empty());
@@ -74,21 +79,27 @@ public class RecordService {
     }
 
 
-    public ResponseEntity<?> getRecordByAccountId(Long accountId) {
+    public ResponseEntity<?> getRecordByAccountId(Long accountId, int page , int size , LocalDate startDate , LocalDate endDate) {
 
         Long userId = AuthUtil.getCurrentUserId();
         accountRepo.findByIdAndUserIdAndStatus(accountId, userId , AccountStatus.ACTIVE.getCode())
             .orElseThrow(() -> new ApiException(Map.of("general", "找不到帳戶")));
 
-        Optional<List<RecordBean>> recordOpt = recordRepo.findByAccountId(accountId);
-        
-        SuccessResponse response;
-        if (recordOpt.isEmpty()) {
-            response = new SuccessResponse(new RecordBean());
-        }else{
-            response = new SuccessResponse(recordOpt.get());
+        Pageable pageable = PageRequest.of(page, size,Sort.by("itemDate").descending());
+        Page<RecordBean> recordPage = null;
+        Long income;
+        Long expense;
+        if (startDate != null && endDate != null) {
+            recordPage = recordRepo.findByAccountIdAndItemDateBetween(accountId, startDate, endDate, pageable);
+            income = recordRepo.sumIncomeBetween(accountId, startDate, endDate);
+            expense = recordRepo.sumExpenseBetween(accountId, startDate, endDate);
+        } else {
+            recordPage = recordRepo.findByAccountId(accountId, pageable);
+            income = recordRepo.sumIncome(accountId);
+            expense = recordRepo.sumExpense(accountId);
         }
-
+        
+        SuccessResponse response = new SuccessResponse(new PageRecordResponse(recordPage,income,expense));
 
         return ResponseEntity.ok(response);
     }
@@ -101,16 +112,10 @@ public class RecordService {
         AccountBean account = accountRepo.findByIdAndUserIdAndStatus(accountId, userId , AccountStatus.ACTIVE.getCode())
             .orElseThrow(() -> new ApiException(Map.of("general", "找不到帳戶")));
 
-        CategoryBean category = categoryRepo.findById(request.getCategory_id())
+        CategoryBean category = categoryRepo.findById(request.getCategoryId())
             .orElseThrow(() -> new ApiException(Map.of("general", "找不到類別")));
 
         RecordBean record = new RecordBean();
-
-        if(category.getType() == -1){
-            account.setInitial_amount(account.getInitial_amount() - request.getItem_price());
-        }else if(category.getType() == 1){
-            account.setInitial_amount(account.getInitial_amount() + request.getItem_price());
-        }
 
         accountRepo.save(account);
 
@@ -118,23 +123,33 @@ public class RecordService {
         user.addRecord(record);
 
         record.setCategory(category);
-        record.setItem_price(request.getItem_price());
-        record.setItem_note(request.getItem_note());
-        record.setItem_date(request.getItem_date());
-        record.setCreated_at(LocalDateTime.now());
-        record.setUpdated_at(LocalDateTime.now());
+        record.setItemPrice(request.getItemPrice());
+        record.setItemNote(request.getItemNote());
+        record.setItemDate(request.getItemDate());
+        record.setCreatedAt(LocalDateTime.now());
+        record.setUpdatedAt(LocalDateTime.now());
         recordRepo.save(record);
 
         return ResponseEntity.ok(new SuccessResponse(null));
     }
 
-    public ResponseEntity<?> recordDelete(long recordId){
-
-        Optional<RecordBean> accountOpt = recordRepo.findById(recordId);
-        if(accountOpt.isEmpty()){
+    public ResponseEntity<?> recordDelete(Long accountId , Long recordId){
+        Long userId = AuthUtil.getCurrentUserId();
+        Optional<RecordBean> recordOpt = recordRepo.findById(recordId);
+        if(recordOpt.isEmpty()){
             throw new ApiException(Map.of("general","刪除失敗"));
         }
-        RecordBean record = accountOpt.get();
+
+        Optional<AccountBean> accountOpt = accountRepo.findByIdAndUserIdAndStatus(accountId, userId, AccountStatus.ACTIVE.getCode());
+        if(accountOpt.isEmpty()){
+            throw new ApiException(Map.of("general","找不到帳戶"));
+        }
+        AccountBean account = accountOpt.get();
+        RecordBean record = recordOpt.get();
+
+        if(!record.getAccount().getId().equals(account.getId())){
+            throw new ApiException(Map.of("general","帳戶不正確"));
+        }
         recordRepo.delete(record);
         return ResponseEntity.ok(new SuccessResponse(null));
     }
